+++ src/makielayout/blocks/legend.jl
@@ -196,6 +196,8 @@ function initialize_block!(leg::Legend,
                 # fill missing entry attributes with those carried by the legend
                 merge!(e.attributes, preset_attrs)
 
+                isnothing(e.label[]) && continue
+
                 # create the label
                 justification = map(leg.labeljustification, e.labelhalign) do lj, lha
                     return lj isa Automatic ? lha : lj
@@ -309,7 +311,7 @@ legendelements(le::LegendElement, legend) = LegendElement[le]
 legendelements(les::AbstractArray{<:LegendElement}, legend) = LegendElement[les...]
 
 
-function LegendEntry(label::AbstractString, contentelements::AbstractArray, legend; kwargs...)
+function LegendEntry(label::Optional{AbstractString}, contentelements::AbstractArray, legend; kwargs...)
     attrs = Attributes(label = label)
 
     kwargattrs = Attributes(kwargs)
@@ -319,7 +321,7 @@ function LegendEntry(label::AbstractString, contentelements::AbstractArray, lege
     LegendEntry(elems, attrs)
 end
 
-function LegendEntry(label::AbstractString, contentelement, legend; kwargs...)
+function LegendEntry(label::Optional{AbstractString}, contentelement, legend; kwargs...)
     attrs = Attributes(label = label)
 
     kwargattrs = Attributes(kwargs)
@@ -464,7 +466,7 @@ one content element. A content element can be an `AbstractPlot`, an array of
 """
 function Legend(fig_or_scene,
         contents::AbstractArray,
-        labels::AbstractArray{<:AbstractString},
+        labels::AbstractArray{<:Optional{AbstractString}},
         title::Optional{<:AbstractString} = nothing;
         kwargs...)